Change positions from double to int to prevent visual glitches caused by decimal points

+++ BPR.Persistence/Models/Position.cs
@@ -2,8 +2,8 @@ namespace BPR.Persistence.Models;
 
 public class Position
 {
-    public double X { get; set; }
-    public double Y { get; set; }
-    public double Height { get; set; }
-    public double Width { get; set; }
+    public int X { get; set; }
+    public int Y { get; set; }
+    public int Height { get; set; }
+    public int Width { get; set; }
 }
